web: replace some more entities and make apostrophes nicer
It turns out that &apos; is equivalent to the vertical
"typewriter" apostrophe '. A true Unicode apostrophe is the
same as a right single quotation mark and can be entered
as Compose->-'. This is all according to the Wikipedia page
on Apostrophe. Sadly, emacs renders the resulting Unicode
characters for me better than Firefox does right now, though
that's presumably an issue of a poor font being chosen.