Change git commit ids to correspond to rebuilt Git repository
I recreated the Git repository with proper committer names, which had
the side effect of changing the SHA1 ids of everything in the repo.
I've found the corresponding commits in the new repo and copied in
the new ids. This may seem a little sketchy, but it requires as much
trust in me as trusting the original ids did.